wrap up rating submission component
